Made new dto pkg in openlibrary pkg
