+++ handler/dto/Book.go
@@ -1,13 +1,5 @@
 package dto
 
-type Book struct {
-	Title      string     `json:"title"`
-	Identifier identifier `json:"identifiers"`
-	Author     []author   `json:"authors"`
-	Cover      cover      `json:"cover"`
-	Year       string     `json:"year"`
-}
-
 type CreateBookRequest struct {
 	ISBN string `json:"ISBN"`
 }
@@ -21,17 +13,4 @@ type BookResponse struct {
 	ID    int    `json:id`
 	Title string `json:"title"`
 	Year  string `json:"year"`
-}
-
-type identifier struct {
-	ISBN10      []string `json:"isbn_10"`
-	ISBN13      []string `json:"isbn_13"`
-	Openlibrary []string `json:"openlibrary"`
-}
-type author struct {
-	Name string `json:"name"`
-}
-
-type cover struct {
-	Url string `json:"small"`
 }
